+++ exporter/opentelemetry-exporter-zipkin/tests/test_zipkin_exporter.py
@@ -223,7 +223,15 @@ class TestZipkinSpanExporter(unittest.TestCase):
                 "annotations": [
                     {
                         "timestamp": event_timestamp // 10 ** 3,
-                        "value": "event0",
+                        "value": json.dumps(
+                            {
+                                "event0": {
+                                    "annotation_bool": True,
+                                    "annotation_string": "annotation_test",
+                                    "key_float": 0.3,
+                                }
+                            }
+                        ),
                     }
                 ],
                 "debug": True,