+++ exporter/opentelemetry-exporter-zipkin/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Zipkin exporter now accepts a ``max_tag_value_length`` attribute to customize the
   maximum allowed size a tag value can have. ([#1151](https://github.com/open-telemetry/opentelemetry-python/pull/1151)) 
+- Fixed OTLP events to Zipkin annotations translation. ([#1161](https://github.com/open-telemetry/opentelemetry-python/pull/1161))
 
 ## Version 0.13b0